requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
et_xmlfile==2.0.0
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
openpyxl==3.1.5
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
pypdfium2==5.1.0
python-barcode==0.16.1
python-dotenv==1.2.1
reportlab==4.4.5
requests==2.32.5
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3

flask-jwt-extended==4.6.0
